requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.1
wincertstore==0.2
